\name{rbaseXT-package}
\alias{rbaseXT-package}
\alias{rbaseXT}
\docType{package}
\title{
\packageTitle{rbaseXT}
}
\description{
\packageDescription{rbaseXT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rbaseXT}
\packageIndices{rbaseXT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rbaseXT}

Maintainer: \packageMaintainer{rbaseXT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
