% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadandinstall.r
\name{loadandinstall}
\alias{loadandinstall}
\title{Load and install packages}
\usage{
loadandinstall(mypkg)
}
\arguments{
\item{mypkg}{String. Single Character String of package name or vector of package names}
}
\description{
Combines the library() and install.packages() comands in one function
}
\examples{
loadandinstall("packagename")
}
\author{
Frederic Schwarzenbacher
}
\keyword{and}
\keyword{install}
\keyword{install.packages#'}
\keyword{library}
\keyword{load}
